requirements.txt
cbor2==5.6.5
crc==7.1.0
pyserial==3.5
